+++ Runtime/Framework/ComposableBehaviour.cs
@@ -173,6 +173,51 @@ namespace DT.UniStart {
       this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
       return action;
     }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction Watch<K>(IEventBus<K> eventBus, K key, UnityAction action) {
+      eventBus.AddListener(key, action);
+      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0> Watch<K, T0>(IEventBus<K> eventBus, K key, UnityAction<T0> action) {
+      eventBus.AddListener(key, action);
+      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0, T1> Watch<K, T0, T1>(IEventBus<K> eventBus, K key, UnityAction<T0, T1> action) {
+      eventBus.AddListener(key, action);
+      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0, T1, T2> Watch<K, T0, T1, T2>(IEventBus<K> eventBus, K key, UnityAction<T0, T1, T2> action) {
+      eventBus.AddListener(key, action);
+      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
+      return action;
+    }
+    /// <summary>
+    /// Watch an event bus for events.
+    /// Remove the listener when the object is destroyed.
+    /// </summary>
+    public UnityAction<T0, T1, T2, T3> Watch<K, T0, T1, T2, T3>(IEventBus<K> eventBus, K key, UnityAction<T0, T1, T2, T3> action) {
+      eventBus.AddListener(key, action);
+      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
+      return action;
+    }
     #endregion
 
     #region Composable Events